libs/cmap: Move into internal
